titanforge_backend/requirements.txt
python-multipart
fastapi
uvicorn[standard]
python-dotenv
chromadb
requests
litellm
gTTS
Pillow
sqlalchemy
psycopg2-binary
redis
stripe
facebook-sdk
ShopifyAPI
apscheduler
pytest
pytest-cov
httpx
Faker
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.0
email-validator
